+++ python-modules.sh
@@ -1,126 +1,114 @@
 package: Python-modules
 version: "1.0"
 requires:
-  - Python
-  - FreeType
+  - "Python:(slc|ubuntu)"
+  - "Python-system:(?!slc.*|ubuntu)"
+  - "FreeType:(?!osx)"
   - libpng
 build_requires:
   - curl
+  - Python-modules-list
+  - alibuild-recipe-tools
 prepend_path:
-  PYTHONPATH: $PYTHON_MODULES_ROOT/lib/python3.6/site-packages:$PYTHONPATH
-prefer_system: (?!slc5)
-prefer_system_check: |
-  python3 -c 'import matplotlib,numpy,scipy,certifi,IPython,ipywidgets,ipykernel,notebook.notebookapp,metakernel,sklearn,six,pymongo,mongoengine,pytest,pylint,yaml'
-  if [ $? -ne 0 ]
-  then
-      printf "Required Python modules are missing. You can install them with pip3:\n  pip3 install matplotlib numpy scipy certifi ipython ipywidgets ipykernel notebook metakernel scikit-learn six pymongo mongoengine pytest pylint pyyaml\n"
-      exit 1
-  fi
+  PATH: "$PYTHON_MODULES_ROOT/share/python-modules/bin"
+  LD_LIBRARY_PATH: "$PYTHON_MODULES_ROOT/share/python-modules/lib"
+  # If we need tensorflow to work on Mac, we must use lib/python$pyver, not lib/python here.
+  PYTHONPATH: $PYTHON_MODULES_ROOT/share/python-modules/lib/python/site-packages
 ---
-#!/bin/bash -ex
-
-if [[ ! $PYTHON_VERSION ]]; then
-  cat <<EoF
-Building our own Python modules.
-If you want to avoid this please install the following modules (pip recommended):
-
-  - matplotlib
-  - numpy
-  - certifi
-  - ipython
-  - ipywidgets
-  - ipykernel
-  - notebook
-  - metakernel
-  - scipy
-  - scikit-learn
-  - six
-  - mock
-  - future
-  - pymongo
-  - mongoengine
-  - pytest
-  - pylint
-  - pyyaml
-EoF
+#!/bin/bash -e
+if [ -n "$VIRTUAL_ENV" ]; then
+  # Once more to get the deactivate
+  . "$VIRTUAL_ENV/bin/activate"
+  deactivate
 fi
-
-# Force pip installation of packages found in current PYTHONPATH
+# A spurious PYTHONPATH can affect later commands
 unset PYTHONPATH
 
-# The X.Y in pythonX.Y
-export PYVER=$(python3 -c 'import distutils.sysconfig; print(distutils.sysconfig.get_python_version())')
+# We use a different INSTALLROOT, so that we can build updatable RPMS which
+# do not conflict with the underlying Python installation.
+PYTHON_MODULES_INSTALLROOT=$INSTALLROOT/share/python-modules
 
-# Install as much as possible with pip. Packages are installed one by one as we
-# are not sure that pip exits with nonzero in case one of the packages failed.
-export PYTHONUSERBASE=$INSTALLROOT
-python3 -m pip install --upgrade pip
-for X in "mock==1.3.0"          \
-         "numpy==1.16.5"        \
-         "certifi==2019.6.16"   \
-         "ipython==5.8.0"       \
-         "ipywidgets==5.2.3"    \
-         "ipykernel==4.10.0"    \
-         "notebook==4.4.1"      \
-         "metakernel==0.24.2"   \
-         "scipy==1.2.2"         \
-         "scikit-learn==0.24.1" \
-         "matplotlib==2.2.4"    \
-         "six"                  \
-         "future"               \
-         "pymongo==3.10.1"      \
-         "pytest==4.6.9"        \
-         "pylint==2.0.1"        \
-         "PyYAML==5.1"          \
-         "requests==2.25.0"     \
-         "mongoengine==0.19.1"
-do
-  python3 -m pip install --user $X
-done
+case $ARCHITECTURE in
+  osx_arm64)
+    # On ARM Macs, we need to install Conda to get Tensorflow with hardware support.
+    # Available version list: https://repo.anaconda.com/miniconda/
+    # The Python version of this Conda env matters! Keep it in sync
+    # with the "PIPXY_REQUIREMENTS_osx_arm64" clause in Python-modules-list.
+    curl -fsSLo miniconda.sh 'https://repo.anaconda.com/miniconda/Miniconda3-py39_23.3.1-0-MacOSX-arm64.sh'
+    bash miniconda.sh -b -p "$PYTHON_MODULES_INSTALLROOT"
+    . "$PYTHON_MODULES_INSTALLROOT/bin/activate"
+    conda install -y -c apple tensorflow-deps ;;
+  *)
+    # On other platforms, just create a plain virtualenv.
+    python3 -m venv "$PYTHON_MODULES_INSTALLROOT"
+    . "$PYTHON_MODULES_INSTALLROOT/bin/activate" ;;
+esac
 
-# for some unknown reason lib directory created with access only for user
-chmod -R 755 $PYTHONUSERBASE/lib
-unset PYTHONUSERBASE
+# Major.minor version of Python
+pyver="$(python3 -c 'import distutils.sysconfig; print(distutils.sysconfig.get_python_version())')"
 
-# Test if matplotlib can be loaded
-env PYTHONPATH="$INSTALLROOT/lib/python3.6/site-packages" python3 -c 'import matplotlib'
+# These are the basic requirements needed for all installation and platform
+# and it should represent the common denominator (working) for all packages/platforms
+echo "$PIP_BASE_REQUIREMENTS" | tr '[:space:]' '\n' > base-requirements.txt
 
-# Remove unneeded stuff
-rm -rvf $INSTALLROOT/share            \
-        $INSTALLROOT/lib/python*/test
-find $INSTALLROOT/lib/python*                                              \
+# PIP*_REQUIREMENTS variables come from python-modules-list.sh.
+case $ARCHITECTURE in
+  slc6_*) echo "$PIP_REQUIREMENTS" ;;
+  *)
+    echo "$PIP_REQUIREMENTS"
+    # Handle special lists for different platforms, e.g. $PIP39_REQUIREMENTS_osx_arm64.
+    this_pyver_requirements_var=PIP${pyver/.}_REQUIREMENTS
+    this_pyver_arch_requirements_var=PIP${pyver/.}_REQUIREMENTS_${ARCHITECTURE//-/_}
+    # Use $PIPxy_REQUIREMENTS_arch if set, falling back to $PIPxy_REQUIREMENTS.
+    echo "${!this_pyver_arch_requirements_var:-${!this_pyver_requirements_var}}" ;;
+esac | tr -s '[:space:]' '\n' > requirements.txt
+
+# Install pinned basic requirements for python infrastructure
+python3 -m pip install -IU -r base-requirements.txt
+
+# FIXME: required because of the newly introduced dependency on scikit-garden requires
+# a numpy to be installed separately
+# See also:
+#   https://github.com/scikit-garden/scikit-garden/issues/23
+python3 -m pip install -IU numpy
+python3 -m pip install -IU -r requirements.txt
+
+# Find the proper Python lib library and export it
+pushd "$PYTHON_MODULES_INSTALLROOT"
+  # let's remove any pre-existent symlinks to have a clean slate
+  [ -h lib64 ] && unlink lib64
+  [ -h lib ]   && unlink lib
+  if [[ -d lib64 ]]; then
+    ln -nfs lib64 lib  # creates lib pointing to lib64
+  elif [[ -d lib ]]; then
+    ln -nfs lib lib64  # creates lib64 pointing to lib
+  fi
+  ln -nfs "python$pyver" lib/python
+popd
+
+# Remove useless stuff
+rm -rvf "$PYTHON_MODULES_INSTALLROOT"/share "$PYTHON_MODULES_INSTALLROOT"/lib/python*/test
+find "$PYTHON_MODULES_INSTALLROOT"/lib/python* \
      -mindepth 2 -maxdepth 2 -type d -and \( -name test -or -name tests \) \
      -exec rm -rvf '{}' \;
 
-# Fix shebangs to point to the correct Python from the runtime environment
-grep -IlRE '#!.*python' $INSTALLROOT/bin | \
-  xargs -n1 perl -p -i -e 's|^#!.*/python|#!/usr/bin/env python|'
-
-# Test whether we can load Python modules (this is not obvious as some of them
-# do not indicate some of their dependencies and break at runtime).
-PYTHONPATH=$INSTALLROOT/lib64/python$PYVER/site-packages:$INSTALLROOT/lib/python$PYVER/site-packages:$PYTHONPATH \
-  python3 -c 'import matplotlib,numpy,scipy,certifi,IPython,ipywidgets,ipykernel,notebook.notebookapp,metakernel,sklearn,six,pymongo,mongoengine,pytest,pylint'
+case $ARCHITECTURE in
+  osx_arm64) ;;
+  *)
+    # Fix shebangs: remove hardcoded Python path
+    find "$PYTHON_MODULES_INSTALLROOT/bin" -type f -exec sed -i.deleteme -e "s|${PYTHON_MODULES_INSTALLROOT}|/usr|;s|python3|env python3|" '{}' \;
+    find "$PYTHON_MODULES_INSTALLROOT/bin" -name '*.deleteme' -delete ;;
+esac
 
 # Modulefile
 MODULEDIR="$INSTALLROOT/etc/modulefiles"
-MODULEFILE="$MODULEDIR/$PKGNAME"
 mkdir -p "$MODULEDIR"
-cat > "$MODULEFILE" <<EoF
-#%Module1.0
-proc ModulesHelp { } {
-  global version
-  puts stderr "ALICE Modulefile for $PKGNAME $PKGVERSION-@@PKGREVISION@$PKGHASH@@"
-}
-set version $PKGVERSION-@@PKGREVISION@$PKGHASH@@
-module-whatis "ALICE Modulefile for $PKGNAME $PKGVERSION-@@PKGREVISION@$PKGHASH@@"
-# Dependencies
-module load BASE/1.0 ${PYTHON_VERSION:+Python/$PYTHON_VERSION-$PYTHON_REVISION} ${ALIEN_RUNTIME_VERSION:+AliEn-Runtime/$ALIEN_RUNTIME_VERSION-$ALIEN_RUNTIME_REVISION}
+alibuild-generate-module > "$MODULEDIR/$PKGNAME"
+cat >> "$MODULEDIR/$PKGNAME" <<EoF
 # Our environment
-setenv PYTHON_MODULES_ROOT \$::env(BASEDIR)/$PKGNAME/\$version
-prepend-path PATH $::env(PYTHON_MODULES_ROOT)/bin
-prepend-path LD_LIBRARY_PATH $::env(PYTHON_MODULES_ROOT)/lib64
-prepend-path LD_LIBRARY_PATH $::env(PYTHON_MODULES_ROOT)/lib
-$([[ ${ARCHITECTURE:0:3} == osx ]] && echo "prepend-path DYLD_LIBRARY_PATH $::env(PYTHON_MODULES_ROOT)/lib64" && \
-                                      echo "prepend-path DYLD_LIBRARY_PATH $::env(PYTHON_MODULES_ROOT)/lib")
-prepend-path PYTHONPATH $::env(PYTHON_MODULES_ROOT)/lib/python$PYVER/site-packages
+set PYTHON_MODULES_ROOT \$::env(BASEDIR)/$PKGNAME/\$version
+prepend-path PATH \$PYTHON_MODULES_ROOT/share/python-modules/bin
+prepend-path LD_LIBRARY_PATH \$PYTHON_MODULES_ROOT/share/python-modules/lib
+# We need to use lib/python$pyver, not lib/python here so that tensorflow-metal works on Mac.
+prepend-path PYTHONPATH \$PYTHON_MODULES_ROOT/share/python-modules/lib/python$pyver/site-packages
 EoF